feat: modernise as per next documentation
