+++ src/components/Nav.tsx
@@ -10,10 +10,8 @@ export default function Nav() {
           xmlns="http://www.w3.org/2000/svg"
         >
           <path d="M13.5 22.1c1.8-7.2 6.3-10.8 13.5-10.8 10.8 0 12.15 8.1 17.55 9.45 3.6.9 6.75-.45 9.45-4.05-1.8 7.2-6.3 10.8-13.5 10.8-10.8 0-12.15-8.1-17.55-9.45-3.6-.9-6.75.45-9.45 4.05zM0 38.3c1.8-7.2 6.3-10.8 13.5-10.8 10.8 0 12.15 8.1 17.55 9.45 3.6.9 6.75-.45 9.45-4.05-1.8 7.2-6.3 10.8-13.5 10.8-10.8 0-12.15-8.1-17.55-9.45-3.6-.9-6.75.45-9.45 4.05z" />
-        </svg>  
-        <span className="font-semibold text-xl tracking-tight">
-          NEWS APP
-        </span>
+        </svg>
+        <span className="font-semibold text-xl tracking-tight">TOP 5 NEWS</span>
       </div>
     </nav>
   );
